feat: implement getPriceLevels

+++ include/Order.hpp
@@ -10,7 +10,7 @@ enum class EventType { ADD_ORDER, CANCEL_ORDER, TRADE };
 using timestamp_t = size_t;
 using eventId_t = size_t;
 using orderId_t = size_t;
-using price_t = size_t;
+using price_t = double;
 using quantity_t = size_t;
 using symbol_t = std::string;
 using type_t = std::string;
